frontend/requirements.txt
streamlit
requests
python-dotenv
bcrypt>=4.0.0
colorama
